working model animation, cleaned up int types

+++ include/resources/resources.h
@@ -19,36 +19,36 @@ struct Texture
 		ID = 0;
 		dim = glm::vec2(1, 1);
 	}
-	Texture(unsigned int ID, glm::vec2 dimentions, std::string path)
+	Texture(size_t ID, glm::vec2 dimentions, std::string path)
 	{
 		this->path = path;
 		this->ID = ID;
 		this->dim = dimentions;
 	}
 	std::string path;
-	unsigned int ID = 0;
+	size_t ID = 0;
 	glm::vec2 dim = glm::vec2(0, 0);
 };
 
 struct Model
 {
 	Model() { this->ID = 1000000;}
-	Model(unsigned int ID)
+	Model(size_t ID)
 	{
 		this->ID = ID;
 	}
-	unsigned int ID;
+	size_t ID;
 };
 
 struct Font
 {
 	Font() { this->ID = 10000000; }
-	Font(unsigned int ID)
+	Font(size_t ID)
 	{
 		this->ID = ID;
 	}
 
-	unsigned int ID;
+	size_t ID;
 };
 
 struct QuadDraw
